+++ package.json
@@ -7,7 +7,7 @@
     "type": "git",
     "url": "https://github.com/manister/netlify-plugin-lazy-load-images.git"
   },
-  "main": "index.js",
+  "main": "out/src/index.js",
   "keywords": [
     "lazy load",
     "netlify",
@@ -16,6 +16,7 @@
   "scripts": {
     "build-test-site": "node build-test-site.js",
     "build-client-js": "node build-client-js.js",
+    "build": "tsc",
     "prepare": "npm run build-client-js"
   },
   "author": "",
@@ -35,6 +36,7 @@
     "@node-minify/babel-minify": "^5.3.0",
     "@node-minify/core": "^5.3.0",
     "babel-preset-env": "^1.7.0",
-    "copy-dir": "^1.3.0"
+    "copy-dir": "^1.3.0",
+    "typescript": "^3.9.5"
   }
 }
