+++ src/lib/imageUrlToPlaceholder.ts
@@ -1,5 +1,4 @@
-
-const getPlaceHolder = require('./getPlaceHolder')
+import getPlaceHolder from './getPlaceHolder'
 
 const imageUrlToPlaceholder = async (imageUrl, paletteSize) => {
   try {
@@ -8,8 +7,7 @@ const imageUrlToPlaceholder = async (imageUrl, paletteSize) => {
     return placeholder
   } catch (e) {
     console.warn("\x1b[31m", `Couldn't get placeholder for ${imageUrl}, ${e}`)
-    return null
+    return ''
   }
 }
-
-module.exports = imageUrlToPlaceholder
+export default imageUrlToPlaceholder
