+++ src/lib/isAbsoluteUrl.ts
@@ -1,6 +1,6 @@
-const isAbsoluteUrl = (url) => {
+const isAbsoluteUrl = (url: string) => {
   const r = new RegExp('^(?:[a-z]+:)?//', 'i');
   return r.test(url)
 }
 
-module.exports = isAbsoluteUrl
+export default isAbsoluteUrl
